requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.1.0
cachetools==5.2.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
frozenlist==1.3.1
fsspec==2022.8.2
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.48.1
idna==3.3
importlib-metadata==4.12.0
Markdown==3.4.1
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.23.3
oauthlib==3.2.1
packaging==21.3
Pillow==9.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.2
pyparsing==3.0.9
python-magic==0.4.17
pytorch-lightning==1.7.3
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scenedetect==0.6.0.3
six==1.16.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.12.1
torchmetrics==0.9.3
torchvision==torchvision-0.13.1
tqdm==4.64.1
typing-extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.8.1
